Revert "feature: implement a `miral::ApplicationSwitcher` as an internal app switching client (#4154)"
This reverts commit 8ef4565d01324e8c5f81f9a14325d41d5e3c6918, reversing
changes made to 67ccd0a1a7e7a4df323d9ff0d89fba1761e1a6bf.
