+++ examples/example-server-lib/decoration_provider.cpp
@@ -47,11 +47,7 @@ struct preferred_codecvt : std::codecvt_byname<wchar_t, char, std::mbstate_t>
 {
     preferred_codecvt() : std::codecvt_byname<wchar_t, char, std::mbstate_t>("C") {}
     ~preferred_codecvt() = default;
-
-    static std::locale::id id;
 };
-
-std::locale::id preferred_codecvt::id;
 
 struct Printer
 {
