+++ examples/example-server-lib/wallpaper_config.cpp
@@ -21,7 +21,7 @@
 
 namespace
 {
-/// Font search logic should be kept in sync with src/common/default_font.cpp
+/// Font search logic should be kept in sync with src/server/shell/decoration/renderer.cpp
 auto default_font() -> std::string
 {
     struct FontPath
