+++ src/CMakeLists.txt
@@ -2,7 +2,7 @@
 set(MIRPLATFORM_ABI 33)
 
 set(MIRAL_VERSION_MAJOR 5)
-set(MIRAL_VERSION_MINOR 6)
+set(MIRAL_VERSION_MINOR 5)
 set(MIRAL_VERSION_PATCH 0)
 set(MIRAL_VERSION ${MIRAL_VERSION_MAJOR}.${MIRAL_VERSION_MINOR}.${MIRAL_VERSION_PATCH})
